following the example from flagger.app
https://docs.flagger.app/tutorials/nginx-progressive-delivery
These are some example resources for a canary with metrics, and an
associated ingress that works on our empfin-projdev cluster.

The nginx ingress is something we already had, with the exception of the
"port" and "scrape" annotations for prometheus, that tell the Prometheus
addon from Flagger where to go for metrics on traffic from the ingress.